src/interview-technical/backend/pyproject.toml
[project]
name = "interview-technical-backend"
version = "0.1.0"
description = "FastAPI backend for multimodal AI technical interviewer"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "dspy-ai>=2.1.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "amazon-transcribe>=0.6.0",
    "websockets>=12.0",
]

[dependency-groups]
dev = ["pytest", "pytest-asyncio"]
